requirements.txt
Pillow
Django
django-cors-headers
djangorestframework